# dist/swagger.yaml
openapi: 3.0.2
servers:
  - url: https://bungou-backend-production.up.railway.app/api/v1
info:
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [http://swagger.io](http://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  version: 1.0.17
  title: Swagger Book Store - OpenAPI 3.0
  termsOfService: 'http://swagger.io/terms/'
  # contact:
  #   email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/book/{id}':
    get:
      tags:
      - book
      summary: Get detail of single book
      description: Get all content about single book by Id
      operationId: getBookById
      parameters:
        - name: id
          in: path
          description: Book ID
          required: true
          schema: 
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Book'
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Invalid BookId supplied
        '404':
          description: Book not found
...
components:
  schemas:
...
    Book:
      properties:
        id:
          type: integer
          format: int64
          example: 10
        title:
          type: string
          example: Normal People
        image:
          type: string
        year:
          type: integer
          example: 2000
        price:
          type: number
          multipleOf: 0.01
        author:
          type: string
          example: Chris Bale
        rating:
          type: number
          multipleOf: 0.1
        publisher:
          type: string
          example: IBM
        length:
          type: integer
          example: 300
        isbn:
          type: string
          example: 12341512523523
        citycountry:
          type: string
          example: London
        catygories:
          type: string
          example: Action
      xml:
        name: book
      type: object
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'https://petstore.swagger.io/oauth/authorize'
          scopes:
            'write:pets': modify pets in your account
            'read:pets': read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header